Hopping on a moving platform always works.  Fixes #3.
If you hop on a moving platform, you used to be in a race
as to who would finish half of their animation.  If the player
got halfway before the moving platform did, everything was fine.
But if the moving platform got halfway first, the player would
be stuck in limbo hanging over empty water.  Fix: remember the
platform you're stepping onto, and if it slid out from underneath
you... well, handle it.